+++ windows/flutter/generated_plugins.cmake
@@ -9,7 +9,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   firebase_auth
   firebase_core
   firebase_storage
-  flutter_webrtc
   geolocator_windows
   record_windows
   share_plus